+++ lib/src/old_widgets/choosers/t_cover_chooser.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:t_widgets/src/downloader/default_downloader.dart';
+import 'package:t_widgets/src/old_widgets/downloader/default_downloader.dart';
 import 'package:t_widgets/t_widgets.dart';
 
 String? initialDirectory;